+++ test/test-dom.js
@@ -5,23 +5,26 @@ const {DOMParser, DOMSerializer, Slice, Fragment, Schema} = require("..")
 
 // declare global: window
 let document = typeof window == "undefined" ? (new (require("jsdom").JSDOM)).window.document : window.document
+const xmlDocument = typeof window == "undefined"
+      ? (new (require("jsdom").JSDOM)("<tag/>", {contentType: "application/xml"})).window.document
+      : window.document
 
 const parser = DOMParser.fromSchema(schema)
 const serializer = DOMSerializer.fromSchema(schema)
 
 describe("DOMParser", () => {
   describe("parse", () => {
-    function domFrom(html) {
-      let dom = document.createElement("div")
+    function domFrom(html, document_ = document) {
+      let dom = document_.createElement("div")
       dom.innerHTML = html
       return dom
     }
 
-    function test(doc, html) {
+    function test(doc, html, document_ = document) {
       return () => {
-        let derivedDOM = document.createElement("div"), schema = doc.type.schema
-        derivedDOM.appendChild(DOMSerializer.fromSchema(schema).serializeFragment(doc.content, {document}))
-        let declaredDOM = domFrom(html)
+        let derivedDOM = document_.createElement("div"), schema = doc.type.schema
+        derivedDOM.appendChild(DOMSerializer.fromSchema(schema).serializeFragment(doc.content, {document: document_}))
+        let declaredDOM = domFrom(html, document_)
 
         ist(derivedDOM.innerHTML, declaredDOM.innerHTML)
         ist(DOMParser.fromSchema(schema).parse(derivedDOM), doc, eq)
@@ -134,6 +137,29 @@ describe("DOMParser", () => {
       let b = builders(markSchema)
       test(b.doc(b.paragraph(b.test("a", b.image({src: "x"}), "b"))),
            "<p><test>a</test><test><img src=\"x\"></test><test>b</test></p>")()
+    })
+
+    it("serializes an element and an attribute with XML namespace", () => {
+      let xmlnsSchema = new Schema({
+        nodes: {
+          doc: { content: "svg*" }, text: {},
+          "svg": {
+            parseDOM: [{tag: "svg", namespace: 'http://www.w3.org/2000/svg'}],
+            group: 'block',
+            toDOM() { return ["http://www.w3.org/2000/svg svg", ["use", { "http://www.w3.org/1999/xlink href": "#svg-id" }]] },
+          },
+        },
+      })
+
+      let b = builders(xmlnsSchema)
+      let d = b.doc(b.svg())
+      test(d, "<svg><use href=\"#svg-id\"></use></svg>", xmlDocument)()
+
+      let dom = xmlDocument.createElement('div')
+      dom.appendChild(DOMSerializer.fromSchema(xmlnsSchema).serializeFragment(d.content, {document: xmlDocument}))
+      ist(dom.querySelector('svg').namespaceURI, 'http://www.w3.org/2000/svg')
+      ist(dom.querySelector('use').namespaceURI, 'http://www.w3.org/2000/svg')
+      ist(dom.querySelector('use').attributes[0].namespaceURI, 'http://www.w3.org/1999/xlink')
     })
 
     function recover(html, doc, options) {
@@ -301,6 +327,12 @@ describe("DOMParser", () => {
        open("<div>foo</div><div><em>bar</em></div>",
             [p("foo"), p(em("bar"))], 1, 1))
 
+    it("will not apply invalid marks to nodes",
+       open("<ul style='font-weight: bold'><li>foo</li></ul>", [ul(li(p(strong("foo"))))], 3, 3))
+
+    it("will apply pending marks from parents to all children",
+       open("<ul style='font-weight: bold'><li>foo</li><li>bar</li></ul>", [ul(li(p(strong("foo"))), li(p(strong("bar"))))], 3, 3))
+
     function find(html, doc) {
       return () => {
         let dom = document.createElement("div")
@@ -402,6 +434,11 @@ describe("DOMParser", () => {
         context: cxDoc.resolve(cxDoc.tag.a)
       }), new Slice(blockquote(hr).content, 0, 0), eq)
     })
+
+    it("can close parent nodes from a rule", () => {
+      let closeParser = new DOMParser(schema, [{tag: "br", closeParent: true}].concat(DOMParser.schemaRules(schema)))
+      ist(closeParser.parse(domFrom("<p>one<br>two</p>")), doc(p("one"), p("two")), eq)
+    })
   })
 
   describe("schemaRules", () => {
@@ -426,10 +463,6 @@ describe("DOMParser", () => {
       })
       ist(DOMParser.schemaRules(schema).map(r => r.tag).join(" "), "em bar foo i")
     })
-
-    const xmlDocument = typeof window == "undefined"
-          ? (new (require("jsdom").JSDOM)("<tag/>", {contentType: "application/xml"})).window.document
-          : window.document
 
     function nsParse(doc, namespace) {
       let schema = new Schema({